Conflicts:
	src/ck_main.c
	src/id_ca.h

+++ src/ck_game.c
@@ -264,7 +264,7 @@ static int ck_cacheCountdownNum, ck_cacheBoxChunksPerPic, ck_cacheBoxChunkCounte
 void CK_BeginCacheBox (const char *title, int numChunks)
 {
 	int totalfree;
-	int w, h;
+	uint16_t w, h;
 
 	// Vanilla keen checks if > 2kb free here
 	// If not, it doesn't cache all of the keen counting down graphics
